be-codepaathshala/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.8.1
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.108
botocore==1.34.108
celery==5.3.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
distro==1.9.0
Django==4.2.11
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-jazzmin==2.6.1
django-storages==1.14.3
django-timezone-field==6.1.0
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
kombu==5.3.5
numpy
openai==1.3.5
packaging==24.0
pandas
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
redis==5.0.3
requests==2.31.0
zoomus==1.2.1
matplotlib
seaborn